+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release Notes
 
 
-## Unreleased
+## v2.3.0
 
 ### Added